+++ CHANGELOG.md
@@ -9,3 +9,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added CLT tasks for Git.
 - Added GitRepository.IsGitHubRepository extension method.
 - Added switches (-major and -minor) for bumping versions when using GitVersionAttribute.
+- Added NukeBuild.InvokedTargets (targets passed from command-line).
+- Added NukeBuild.ExecutingTargets (targets that will be executed).
+- Deprecated -Target parameter in favor of passing targets as first argument to the bootstrapping scripts.